secure-llm-poc-v2/requirements.txt
python-dotenv
openai
langchain
guardrails-ai
promptfoo
rich
pydantic